Tag images appropriately when making snapshot
Uses variables from .env for naming snapshot image

Whe `make snapshot-image` is invoked, the snapshot image will
be given a unique tag name, .env will be updated, and the
docker-compowe.yml file will be re-generated so that it contains the new
tag.  That way, a developer can choose to commit the new snapshot they
built easily (by committing .env), and a docker-compose down -v will
result in restarting to the state of their last snapshot.