Add `snapshot-push` to Makefile
We add a `make snapshot-push` target that invokes `docker push` with the
appropriate arguments pushing the snapshot image to the container
registry.

The previous instructions presumed you could do `docker-compose push` in
order to push snapshot images to the container registry.  However
we removed the build instruction from docker-compose in 9fc6c26.  The
subtle consequence of this is that `docker-compose push` will silently
refuse to push anything, since the lack of a build instruction
implicitly tells it that there is nothing to push.  Hence, we add a Make
target and perform the push via `docker push` rather than
`docker-compose push`.