Updates migrations with using field_unique_id as default for parse_entity_lookups (#204)
Makes it so any taxonomy field that references a user managed taxonomy (which is anything but Media Use, Display and Model) uses the quad formatting and has field_unique_id as the default field for the quad. 
Updates CSV files accordingly. 

Fixes util.js doMigration code to not fail when the drupal update messages on the page.