View for JScholarship Redirects
This adds a View for exporting a file that is similar
to the format used by Apache's RewriteMap.

To access the endpoint for this view:

```
curl -X GET https://test.digital.library.jhu.edu/jscholarship_redirects_rest\?query=\(ss_type\:islandora_object\)\&items_per_page\=500\&offset\=0 > out.csv
```

That will only get the first 500 items. You'll want to use the offset parameter
to go through until you have exported metadata for all the items in the repo.

This would preferably be done in a script.

The resulting file will look like:

```
dspace_identifier "Citable URL"
https://jscholarship.library.jhu.edu/handle/1774.2/57496 https://test.digital.library.jhu.edu/node/5321
```

To use the resulting file as a RewriteMap the headers should be removed and the root URL
should be removed from the jscholarship URLS:

```
/handle/1774.2/57496 https://test.digital.library.jhu.edu/node/5321
```

Then in the JScholarship Apache config (after moving the file to `/etc/httpd/conf/redirects.txt`):

```
  RewriteEngine On
  RewriteMap redirects "txt:/etc/httpd/conf/redirects.txt"
  RewriteCond ${redirects:$1} !=""
  RewriteRule ^(.*)$ ${redirects:$1} [redirect=permanent,last]
```

Use space as separator and ensure only items with dpsace URLs are in the View